feat: 整体优化，适配 v3 版本

+++ lazycatapp/src/main/java/xyz/mxue/lazycatapp/model/response/app/Rating.java
@@ -11,7 +11,7 @@ public class Rating {
     /**
      * 分数
      */
-    private int score;
+    private Double score;
 
     /**
      * 评分统计
